Split MappingEditor into its own file for better code org
